refactor: reorganize directory structure
