+++ src/lib/pose.ts
@@ -1,0 +1,5 @@
+import { Vector } from "./vector";
+
+type Pose = { vector: Vector; angle: number };
+
+export { type Pose };
