+++ src/lib/vector.test.ts
